+++ src/main/java/blackjack/dto/GamerDto.java
@@ -4,8 +4,6 @@ import java.util.List;
 
 import blackjack.domain.gamer.Gamer;
 
-// TODO: 재사용 가능한 DTO로 수정하기
-// TODO: DTO 패키지 분리하기
 public record GamerDto(
     String name,
     List<CardDto> cards
