+++ src/main/java/blackjack/service/BlackjackService.java
@@ -7,11 +7,11 @@ import blackjack.domain.deck.Deck;
 import blackjack.domain.deck.RandomCardStrategy;
 import blackjack.domain.gamer.Dealer;
 import blackjack.domain.gamer.Player;
-import blackjack.dto.FinalResultResponseDto;
 import blackjack.dto.GamerDto;
-import blackjack.dto.NamesRequestDto;
-import blackjack.dto.RoundResultsResponseDto;
-import blackjack.dto.StartingCardsResponseDto;
+import blackjack.dto.request.NamesRequestDto;
+import blackjack.dto.response.FinalResultResponseDto;
+import blackjack.dto.response.RoundResultsResponseDto;
+import blackjack.dto.response.StartingCardsResponseDto;
 
 public class BlackjackService {
